https://sites.google.com/paruluniversity.ac.in/tableauprojectoncovid-19?usp=share_link
As a data analyst with Tableau expertise, I have created a comprehensive project on the global spread 
of COVID-19. Utilizing various sources of data, I have created interactive dashboards and visualizations 
to track the progression of the pandemic. 

The project showcases the amount of people who are fully, partially and not vaccinated .

The dashboards enable users to explore the data in an intuitive and visually appealing way, making it easier to understand the complex and rapidly evolving situation. Through this project, I have demonstrated my ability to analyze and present complex data in a meaningful and impactful way, highlighting the importance of data-driven decision making in today's world.